검색 바텀시트 작업

+++ app/src/main/java/com/example/booksapp/core/domain/model/Book.kt
@@ -7,7 +7,7 @@ data class Book(
     val date: String,
     val priceStandard: Int,
     val priceSales: Int,
-    val discountRate: Int,
+    val discountRate: Double,
     val saleStatus: String,
     val mileage: String,
     val mileageRate: String,
